Assignment 3
Since we will be working with programming languages like C++, Python, and Java in the future, I think that learning these abilities is essential to succeeding in the field of computer science. Soft qualities like empathy, flexibility, and teamwork are also necessary for me to operate well in a variety of settings. I must work on a lot of group projects as these will help me network with my professors and peers.